Move cluster deletion into pkg/clusterdeletion
