+++ api/pkg/clusterdeletion/OWNERS
@@ -1,0 +1,13 @@
+# See the OWNERS docs: https://git.k8s.io/community/contributors/guide/owners.md
+
+approvers:
+  - team-lifecycle
+
+reviewers:
+  - team-lifecycle
+
+labels:
+  - team/lifecycle
+
+options:
+  no_parent_owners: true
