+++ api/pkg/clusterdeletion/metrics.go
@@ -1,4 +1,4 @@
-package deletion
+package clusterdeletion
 
 import (
 	"sync"
